Refacto settings dialog + update readme
