+++ src/components/settings-dialog/appearance-section.ts
@@ -1,0 +1,57 @@
+import { consume } from "@lit/context";
+import { LitElement, html } from "lit";
+import { customElement, state } from "lit/decorators.js";
+
+import type { LocalizeFunc } from "../../common/localize.js";
+import { localizeContext } from "../../context/index.js";
+import { espHomeStyles } from "../../styles/shared.js";
+import { settingsRowStyles, settingsSharedStyles } from "./shared-styles.js";
+
+import "@home-assistant/webawesome/dist/components/option/option.js";
+import "@home-assistant/webawesome/dist/components/select/select.js";
+
+@customElement("esphome-settings-appearance")
+export class ESPHomeSettingsAppearance extends LitElement {
+  @consume({ context: localizeContext, subscribe: true })
+  @state()
+  private _localize: LocalizeFunc = (key) => key;
+
+  @state()
+  private _theme: string = localStorage.getItem("esphome-theme") ?? "system";
+
+  static styles = [espHomeStyles, settingsSharedStyles, settingsRowStyles];
+
+  protected render() {
+    return html`
+      <div class="row">
+        <div class="row-label">
+          <span class="row-title">${this._localize("layout.theme")}</span>
+          <span class="row-desc">${this._localize("settings.theme_desc")}</span>
+        </div>
+        <wa-select value=${this._theme} @change=${this._onChange}>
+          <wa-option value="light">${this._localize("layout.theme_light")}</wa-option>
+          <wa-option value="dark">${this._localize("layout.theme_dark")}</wa-option>
+          <wa-option value="system">${this._localize("layout.theme_system")}</wa-option>
+        </wa-select>
+      </div>
+    `;
+  }
+
+  private _onChange(e: Event) {
+    const theme = (e.target as HTMLSelectElement).value;
+    this._theme = theme;
+    this.dispatchEvent(
+      new CustomEvent("set-theme", {
+        detail: theme,
+        bubbles: true,
+        composed: true,
+      }),
+    );
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    "esphome-settings-appearance": ESPHomeSettingsAppearance;
+  }
+}
